fix loc.start.column = -1 in parser snapshot
